ui_playground: rename module to `viewlib`, move components
